BBProj/requirements.txt
argon2-cffi==20.1.0
asgiref==3.5.2
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==8.0.4
colorama==0.4.3
cryptography==36.0.2
cx-Oracle==8.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==4.0.6
djangorestframework==3.13.1
entrypoints==0.3
Flask==2.0.3
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==2.1.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==2.1.1
matplotlib==3.3.3
mistune==0.8.4
mysql-connector-python==8.0.24
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.5
packaging==20.4
pandas==1.2.1
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycparser==2.20
Pygments==2.7.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.2
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==5.0.4
tzdata==2022.1
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.1
